split demo into sudmodules
